Fix compiler warnings

+++ src/batch/extract_corr_matrix.cc
@@ -56,7 +56,7 @@ int main(int argc, char* argv[])
     // Iterate over each file, and label each correlation matrix block with the file
     // name
     bool is_header_written = false; // only write header once
-    int num_params = 0; // store number of parameters to ensure other files match
+    long unsigned int num_params = 0; // store number of parameters to ensure other files match
     for (const std::string &file : file_vector)
     {
         std::cout << "Analyzing File: " << file << "\n";
@@ -104,7 +104,7 @@ int main(int argc, char* argv[])
             exit(1);
         }
 
-        for (int row = 0; row < cov_matrix.size(); row++)
+        for (long unsigned int row = 0; row < cov_matrix.size(); row++)
         {
             const std::string row_par = results.parNameList()[row];
             const double row_par_error = results.parError(row_par);
@@ -116,7 +116,7 @@ int main(int argc, char* argv[])
             csv_data << file << ",";
             csv_data << row_par;
 
-            for (int col = 0; col < cov_matrix[row].size(); col++)
+            for (long unsigned int col = 0; col < cov_matrix[row].size(); col++)
             {
                 const std::string col_par = results.parNameList()[col];
                 const double col_par_error = results.parError(col_par);
